chore: Move ConnectionManager to backend folder and update imports
