+++ backend/__init__.py
@@ -1,1 +1,2 @@
 from .search import SearchQuery  # noqa: F401
+from .connection_manager import ConnectionManager  # noqa: F401
