+++ backend/connection_manager.py
